requirements.txt
googleapis-common-protos==1.52.0
grpcio==1.32.0
grpcio-tools==1.32.0
protobuf==3.13.0
six==1.15.0
